Value is now in double

+++ analyzer/src/analyzer/AnalysisMethod.java
@@ -13,7 +13,7 @@ public interface AnalysisMethod {
 	 * @return the value of the method on the specific stock. If it indicates wether the stock is oerbought or not its an 
 	 * integer between 0 and 100.
 	 */
-	public int value();
+	public double value();
 	/**
 	 * 
 	 * @return a set of curves that represents the graph of the method, some methods may use only one curve
